requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-extensions
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
python-dotenv==0.10.3
Werkzeug==0.15.4
whitenoise==5.2.0
picamera